# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsentCreate": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "enum": [
              "ALL",
              "EMAIL",
              "SMS",
              "VOICE",
              "SOCIAL"
            ],
            "title": "Channel",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "identity": {
            "$ref": "#/components/schemas/IdentityReference"
          },
          "lawful_basis": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Lawful Basis",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "source": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Source",
            "type": "string"
          },
          "state": {
            "enum": [
              "GRANTED",
              "DENIED",
              "REVOKED"
            ],
            "title": "State",
            "type": "string"
          },
          "topic": {
            "pattern": "^[a-z0-9][a-z0-9._-]{0,119}$",
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "identity",
          "topic",
          "channel",
          "state",
          "lawful_basis",
          "source",
          "occurred_at"
        ],
        "title": "ConsentCreate",
        "type": "object"
      },
      "ConsentOut": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_hash": {
            "title": "Evidence Hash",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "identity_reference": {
            "title": "Identity Reference",
            "type": "string"
          },
          "lawful_basis": {
            "title": "Lawful Basis",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "profile_id",
          "identity_reference",
          "topic",
          "channel",
          "state",
          "lawful_basis",
          "source",
          "evidence_hash",
          "occurred_at",
          "created_at"
        ],
        "title": "ConsentOut",
        "type": "object"
      },
      "EffectivePreferenceOut": {
        "additionalProperties": false,
        "properties": {
          "allowed": {
            "title": "Allowed",
            "type": "boolean"
          },
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "consent_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent State"
          },
          "identity_reference": {
            "title": "Identity Reference",
            "type": "string"
          },
          "preference_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference State"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "identity_reference",
          "topic",
          "channel",
          "allowed",
          "reason",
          "consent_state",
          "preference_state"
        ],
        "title": "EffectivePreferenceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdentityReference": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "EMAIL",
              "PHONE",
              "EXTERNAL"
            ],
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "maxLength": 320,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "title": "IdentityReference",
        "type": "object"
      },
...
      "PreferenceOut": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "identity_reference": {
            "title": "Identity Reference",
            "type": "string"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "profile_id",
          "identity_reference",
          "topic",
          "channel",
          "state",
          "frequency",
          "source",
          "version",
          "updated_at"
        ],
        "title": "PreferenceOut",
        "type": "object"
      },
      "PreferenceUpsert": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "enum": [
              "ALL",
              "EMAIL",
              "SMS",
              "VOICE",
              "SOCIAL"
            ],
            "title": "Channel",
            "type": "string"
          },
          "expected_version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "frequency": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency"
          },
          "identity": {
            "$ref": "#/components/schemas/IdentityReference"
          },
          "source": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Source",
            "type": "string"
          },
          "state": {
            "enum": [
              "SUBSCRIBED",
              "UNSUBSCRIBED"
            ],
            "title": "State",
            "type": "string"
          },
          "topic": {
            "pattern": "^[a-z0-9][a-z0-9._-]{0,119}$",
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "identity",
          "topic",
          "channel",
          "state",
          "source"
        ],
        "title": "PreferenceUpsert",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Shared tenant, profile, identity, consent, preference, billing, usage, and entitlement authority. Effectful cross-system calls remain governed by Codestra Middleware.",
    "title": "Codestra Foundation API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tenants/{tenant_id}/communications/consents": {
      "get": {
        "operationId": "list_consents_v1_tenants__tenant_id__communications_consents_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "identity_kind",
            "required": true,
            "schema": {
              "pattern": "^(EMAIL|PHONE|EXTERNAL)$",
              "title": "Identity Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "identity_value",
            "required": true,
            "schema": {
              "maxLength": 320,
              "minLength": 1,
              "title": "Identity Value",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsentOut"
                  },
                  "title": "Response List Consents V1 Tenants  Tenant Id  Communications Consents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Consents",
        "tags": [
          "consent-preferences"
        ]
      },
      "post": {
        "operationId": "append_consent_v1_tenants__tenant_id__communications_consents_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 240,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 8,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Consent",
        "tags": [
          "consent-preferences"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/communications/effective": {
      "get": {
        "operationId": "effective_preference_v1_tenants__tenant_id__communications_effective_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "identity_kind",
            "required": true,
            "schema": {
              "pattern": "^(EMAIL|PHONE|EXTERNAL)$",
              "title": "Identity Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "identity_value",
            "required": true,
            "schema": {
              "maxLength": 320,
              "minLength": 1,
              "title": "Identity Value",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9._-]{0,119}$",
              "title": "Topic",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "channel",
            "required": true,
            "schema": {
              "pattern": "^(EMAIL|SMS|VOICE|SOCIAL)$",
              "title": "Channel",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectivePreferenceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Effective Preference",
        "tags": [
          "consent-preferences"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/communications/preferences": {
      "put": {
        "operationId": "upsert_preference_v1_tenants__tenant_id__communications_preferences_put",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 240,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 8,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Preference",
        "tags": [
          "consent-preferences"
        ]
      }
    },
...
  }
}